mejora al encontrar expansiones para calcular puntos.

+++ tsinc/utils.py
@@ -86,7 +86,7 @@ def print_column(sheet,data,row,start_column,config=None):
         if config.get('fill', False):
             cell.fill = PatternFill(start_color=config.get('fill_color', "FFFFFF"), end_color=config.get('fill_color', "FFFFFF"), fill_type="solid")
 
-def closest_match(input_one,input_two):
+def closest_match(input_one,input_two):# elige la combinacion que mas se acerque a los puntos generados en los tableros de control
 
     def generar_combinaciones(num_ui, num_co, bo, bi, ao):
         combinaciones = []
@@ -166,43 +166,52 @@ def find_expansion(total_points,controller_and_expansions,expansions):
     
     expansion_choice = None
     shoter_distance = float('inf')
-    points = [ item['points'] for item in controller_and_expansions]
-    add_p = add_points(*points)
-    for expansion in expansions:
-        cller_exp_point = add_points(expansion['points'],add_p) 
+    points = [ item['points'] for item in controller_and_expansions]# agrega los puntos de todas las expansiones que se van agregando hasta el momento
+  
+    add_p = add_points(*points) # suma todos los puntos de los modulos que se han agregado
+
+    for expansion in expansions: #Recorremos todas las lista expansiones disponibles que son las expansiones filtradas 
+        cller_exp_point = add_points(expansion['points'],add_p) # va sumando los puntos de las expansiones hasta el momento con cada una de las expansiones disponibles
         # print(f"{expansion['points']}+{add_p} result:{cller_exp_point}")
-        new_cller_exp_point = closest_match(cller_exp_point,total_points)
-        expansion['ce_basic_points'] = new_cller_exp_point
-        distance = calc_distance(new_cller_exp_point, total_points)
+        new_cller_exp_point = closest_match(cller_exp_point,total_points)# de la anterior combinacion hace un close_match para encontrar la combinación mas cercana para convertirlo a los puntos basicos que son [ EA, ED, SA, SD]
+       
+        expansion['ce_basic_points'] = new_cller_exp_point # guarda los puntos basicos en cada expansion
+        distance = calc_distance(new_cller_exp_point, total_points) #calcula la distancia mas cercana con el total de puntos proyectados para elegir la expansion
         #print(total_points, f"todo:{new_cller_exp_point}{expansion['model']}:{distance}")
         if distance < shoter_distance:
             shoter_distance = distance
             expansion_choice = expansion
-    return expansion_choice
+    return expansion_choice # devuelve la expansion elegida
         
 def calc_controllers(item_points, type_controller):
     
-    def iterator_function(total_points,filtered_expansions):
-        expansion = find_expansion(total_points,controller_and_expansions,filtered_expansions)
+    def iterator_function(total_points,filtered_expansions):#
+        expansion = find_expansion(total_points,controller_and_expansions,filtered_expansions)#Encuentra las expansiones por distancia 
         controller_and_expansions.append(expansion)
-        new_points = subtract_points(total_points, expansion['ce_basic_points'])
+        new_points = subtract_points(total_points, expansion['ce_basic_points'])# resta los puntos 
         return new_points
     
-    def filter_expansions(total_points,expansions):
-        prom = 10
-        total_points = clean_points(total_points)
-        reduce_points = functools.reduce(lambda x,y:x+y,total_points)
-        if reduce_points <= prom:
-            filtered_expansions = [expansion for expansion in expansions if expansion['total_points'] <= prom and expansion['total_points']>0]
-        elif reduce_points > prom:
-            filtered_expansions = [expansion for expansion in expansions if expansion['total_points'] > prom ]
+    def filter_expansions(total_points,expansions): # funcion que se encarga de filtrar las expansiones segun los puntos que hagan falta por cubrir
+    
+        filtered_expansions = []
+        for expansion in expansions:
+        # Condiciones para agregar a la lista filtrada
+            if (
+                (total_points[0] > 0 and expansion["points"][0] != 0) or
+                (total_points[1] > 0 and expansion["points"][0] != 0 and expansion["points"][1] != 0) or
+                (total_points[2] > 0 and expansion["points"][2] != 0 and expansion["points"][3] != 0) or
+                (total_points[3] > 0 and expansion["points"][3] != 0 and expansion["points"][4] != 0)
+            ):
+                filtered_expansions.append(expansion)
+
+        
         return filtered_expansions
         
     
-    def add_expansion_js(total_points,expansions):
+    def add_expansion_js(total_points,expansions): # recibe los punto totales generados de cada tablero y expansiones en diccionario con sus puntos
         stop=0
-        filtered_expansions = filter_expansions(total_points, expansions)
-        new_points = iterator_function(total_points,filtered_expansions)
+        filtered_expansions = filter_expansions(total_points, expansions) # filtra la expansiones con el total de puntos
+        new_points = iterator_function(total_points,filtered_expansions) # calcula la ditancia del controlador y expansione que se van agregando
         while any( point > 0 for point in new_points):
             stop +=1
             filtered_expansions = filter_expansions(new_points, expansions)
@@ -237,12 +246,12 @@ def calc_controllers(item_points, type_controller):
     def find_cllers_exp(cller_type):
          
         if cller_type == "JCI FACILITY EXPLORER":
-            controllers = Product.objects.filter(code="C-FE")
-            expansions = Product.objects.filter(code='E-FE')
+            controllers = Product.objects.filter(code__icontains="C-FE")
+            expansions = Product.objects.filter(code__icontains='E-FE')
         else:
-            controllers = Product.objects.filter(code="C-M")
-            expansions = Product.objects.filter(code='E-M')
-        # print(expansions)
+            controllers = Product.objects.filter(code__icontains="C-M")
+            expansions = Product.objects.filter(code__icontains='E-M')
+        
        
         controllers_js = [{
             'is_controller':True,
@@ -255,7 +264,7 @@ def calc_controllers(item_points, type_controller):
             'is_controller':False,
             'model':expansion.model,
             'points':sort_list_point(expansion.point.split("-")),
-            'total_points': functools.reduce(lambda x,y: x+y,sort_list_point(expansion.point.split("-")))
+            'total_points': functools.reduce(lambda x,y: x+y,sort_list_point(expansion.point.split("-"))),
         }
         for expansion in expansions  
         ]   
@@ -273,7 +282,8 @@ def calc_controllers(item_points, type_controller):
         expansions_lg = [{
             'is_controller':False,
             'model':expansion.model,
-            'points':sort_list_point(expansion.point.split("-")) 
+            'points':sort_list_point(expansion.point.split("-")), 
+            
         }
         for expansion in expansions   
         ]
@@ -300,10 +310,15 @@ def calc_controllers(item_points, type_controller):
             if point_quantity > 0 and point_quantity < 18:
 
                 modelos_deseados = {"F4-CGM04060-0", "M4-CGM04060-0"}
+
+
                 
                 # Filtra los controladores utilizando una comprensión de lista
                 controller = next((controller for controller in controllers_js if controller['model'] in modelos_deseados), None)
 
+                if not controller:
+                    raise ProductNotFoundError(f"No se ha podido encontrar alguno de estos controladores: {modelos_deseados} en la base de datos.") 
+                
                 new_controller_point = closest_match(controller['points'],item_points)
                 
                 #print(f'puntos controlador:{new_controller_point}')
@@ -319,6 +334,11 @@ def calc_controllers(item_points, type_controller):
                 
                 # Filtra los controladores utilizando una comprensión de lista
                 controller = next((controller for controller in controllers_js if controller['model'] in modelos_deseados), None)
+
+                if not controller:
+                    raise ProductNotFoundError(f"No se ha podido encontrar alguno de estos controladores: {modelos_deseados} en la base de datos.") 
+
+
                 
                 new_controller_point = closest_match(controller['points'],item_points)
                 
@@ -545,6 +565,10 @@ def calc_chest(c_quantity):
         chest = Product.objects.filter(code="CO5050").first()
     return chest    
     
+class ProductNotFoundError(Exception):
+    def __init__(self, message="Product not found"):
+        self.message = message
+        super().__init__(self.message)
 
 
 def print_points_cller_exp_trf(sheet,points,cllers,sheet_model,project,modify=False):
@@ -753,17 +777,24 @@ def print_points_cller_exp_trf(sheet,points,cllers,sheet_model,project,modify=Fa
     quantity_trf = [quantity,trf]
     
     if not modify:
-        trf = Product.objects.filter(model = trf).first()
+        trf_ = Product.objects.filter(model = trf).first()
+        if not trf_:
+            raise ProductNotFoundError(f"No se ha podido encontrar un transfomador: {trf} en la base de datos.") 
         tab = Tabs.objects.filter(tab_name = sheet.title, project=project).first()
         for i in range(1,quantity+1):
-            Divice.objects.create(model=trf.model,brand = trf.brand, tab = tab, project = project)
+            Divice.objects.create(model=trf_.model,brand = trf_.brand, tab = tab, project = project)
     else:
-        trf = Product.objects.filter(model = trf).first()
+        
+        trf_ = Product.objects.filter(model = trf).first()
+        
+        if not trf_:
+            raise ProductNotFoundError(f"No se ha podido encontrar un transfomador: {trf} en la base de datos.") 
+    
         tab = Tabs.objects.filter(tab_name = sheet.title, project= project).first()
-        divice = Divice.objects.filter(model=trf.model, tab = tab, project = project).all()
+        divice = Divice.objects.filter(model=trf_.model, tab = tab, project = project).all()
         divice.delete()
         for i in range(1,quantity+1):
-            Divice.objects.create(model=trf.model,brand = trf.brand, tab = tab, project = project)
+            Divice.objects.create(model=trf_.model,brand = trf_.brand, tab = tab, project = project)
     
     print_column(sheet,quantity_trf,row3+4,11,{
         'font':False,
